requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
absl-py==2.1.0
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asdl==0.1.5
attrs==24.2.0
auto_gptq==0.7.1
babel==2.16.0
blinker==1.9.0
blis==0.7.11
bpemb==0.3.6
captum==0.7.0
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
corenlp
cryptography==3.4.8
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.10
dacite==1.8.1
datasets==3.1.0
#dbus-python==1.2.18
dill==0.3.8
distro==1.7.0
docker-pycreds==0.4.0
docopt==0.6.2
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl
entmax==1.3
et_xmlfile==2.0.0
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.2.0
gekko==1.2.1
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.68.1
hkkang_utils==0.2.45
httplib2==0.20.2
huggingface-hub==0.26.5
hydra-core
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.2
jpype1==1.5.1
jsonnet==0.20.0
keyring==23.5.0
kiwisolver==1.4.7
konlpy==0.6.0
langcodes==3.5.0
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
meson==1.6.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
networkx==3.2.1
ninja==1.11.1.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
peft==0.14.0
pglast==7.1
pillow==11.0.0
platformdirs==4.3.6
preshed==3.0.9
propcache==0.2.1
protobuf==3.20.0
psutil==6.1.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pyarrow==18.1.0
#pycairo==1.27.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
#PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
records==0.6.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
safetensors==0.4.5
scipy==1.13.1
SecretStorage==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
slack_sdk==3.33.5
smart-open==7.0.5
smmap==5.0.1
spacy==3.7.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stanfordnlp==0.2.0
sympy==1.13.1
tablib==3.7.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
text-generation==0.7.0
thinc==8.2.5
tokenizers==0.21.0
torch==2.2.0+cu121
torchdata==0.7.1
torchtext==0.17.0+cpu
tqdm==4.66.5
transformers==4.47.0
triton==2.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==1.26.13
wadllib==1.3.6
waitress==3.0.2
wandb==0.19.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==1.0.0